refactor: memory_tools.py
